+++ cli/tests/extensions.rs
@@ -1,5 +1,10 @@
 #![cfg(feature = "extensions")]
 
+#[path = "extensions/module_loader.rs"]
+mod module_loader;
+#[path = "extensions/permissions.rs"]
+mod permissions;
+
 use std::convert::TryFrom;
 use std::env;
 use std::path::{Path, PathBuf};
@@ -278,256 +283,6 @@ fn conflicting_extension_name_is_filtered() {
 
     let output = std::str::from_utf8(&cmd.get_output().stderr).unwrap();
     assert!(output.contains("extension was filtered out"));
-}
-
-mod permissions {
-    use std::ffi::OsStr;
-
-    use assert_cmd::assert::Assert;
-    use predicates::prelude::*;
-
-    use super::*;
-
-    struct TestCli {
-        tempdir: TempDir,
-        cwd: Option<PathBuf>,
-    }
-
-    impl Default for TestCli {
-        fn default() -> Self {
-            Self { tempdir: TempDir::new().unwrap(), cwd: None }
-        }
-    }
-
-    impl TestCli {
-        fn new() -> Self {
-            Default::default()
-        }
-
-        fn cwd(mut self, cwd: PathBuf) -> Self {
-            self.cwd = Some(cwd);
-            self
-        }
-
-        fn install_extension(&self, path: &Path) -> Assert {
-            self.run(&["extension", "install", "-y", &path.to_string_lossy()])
-        }
-
-        fn run<S: AsRef<str> + AsRef<OsStr>>(&self, args: &[S]) -> Assert {
-            let mut cmd = Command::cargo_bin("phylum").unwrap();
-
-            cmd.env("XDG_DATA_HOME", self.tempdir.path()).args(args);
-
-            if let Some(cwd) = self.cwd.as_ref() {
-                cmd.current_dir(cwd);
-            }
-
-            cmd.assert()
-        }
-    }
-
-    #[test]
-    fn permission_dialog_is_shown_without_yes_flag() {
-        let test_cli = TestCli::new().cwd(fixtures_path().join("permissions"));
-
-        test_cli
-            .run(&[
-                "extension",
-                "install",
-                &fixtures_path().join("permissions").join("correct-read-perms").to_string_lossy(),
-            ])
-            .failure()
-            .stderr(predicate::str::contains("Can't ask for permissions"));
-    }
-
-    #[test]
-    fn correct_read_permission_successful_install_and_run() {
-        let test_cli = TestCli::new().cwd(fixtures_path().join("permissions"));
-
-        test_cli
-            .install_extension(&fixtures_path().join("permissions").join("correct-read-perms"))
-            .success();
-
-        test_cli
-            .run(&["correct-read-perms"])
-            .success()
-            .stdout(predicate::str::contains("await Deno.readFile"));
-    }
-
-    #[test]
-    fn incorrect_read_permission_unsuccessful_run() {
-        let test_cli = TestCli::new().cwd(fixtures_path().join("permissions"));
-
-        test_cli
-            .install_extension(&fixtures_path().join("permissions").join("incorrect-read-perms"))
-            .success();
-
-        test_cli
-            .run(&["incorrect-read-perms"])
-            .failure()
-            .stderr(predicate::str::contains("Error: Requires read access"));
-    }
-
-    #[test]
-    fn correct_net_permission_successful_install_and_run() {
-        let test_cli = TestCli::new().cwd(fixtures_path().join("permissions"));
-
-        test_cli
-            .install_extension(&fixtures_path().join("permissions").join("correct-net-perms"))
-            .success();
-
-        test_cli
-            .run(&["correct-net-perms"])
-            .success()
-            .stdout(predicate::str::contains("upload_url"));
-    }
-
-    #[test]
-    fn incorrect_net_permission_unsuccessful_run() {
-        let test_cli = TestCli::new().cwd(fixtures_path().join("permissions"));
-
-        test_cli
-            .install_extension(&fixtures_path().join("permissions").join("incorrect-net-perms"))
-            .success();
-
-        test_cli
-            .run(&["incorrect-net-perms"])
-            .failure()
-            .stderr(predicate::str::contains(r#"Error: Requires net access to "api.github.com""#));
-    }
-
-    #[test]
-    fn correct_run_permission_successful_install_and_run() {
-        let test_cli = TestCli::new().cwd(fixtures_path().join("permissions"));
-
-        test_cli
-            .install_extension(&fixtures_path().join("permissions").join("correct-run-perms"))
-            .success();
-
-        test_cli.run(&["correct-run-perms"]).success().stdout(predicate::str::contains("install"));
-    }
-}
-
-// Test that the rules of the module loader are obeyed:
-// - Both TypeScript and JavaScript are supported.
-// - Files under $XDG_DATA_HOME/phylum/extensions may be imported.
-// - Symlinks are not allowed.
-// - Remote URLs under https://deno.land are supported -- i.e., the Deno's
-//   standard library.
-// - No other URLs are supported.
-//   - We explicitly test that a https:// url which is not under `deno.land` is
-//     rejected.
-//   - We explicitly test that a directory traversal attempt is rejected.
-//
-// These tests are based on the fixtures under
-// `fixtures/module-import-extension`.
-mod module_loader_tests {
-    use super::*;
-
-    // The fixture for this test requires one local .ts file, one local .js file,
-    // and one file from Deno's standard library.
-    #[test]
-    fn good_module_loads_successfully() {
-        let tempdir = TempDir::new().unwrap();
-
-        Command::cargo_bin("phylum")
-            .unwrap()
-            .env("XDG_DATA_HOME", tempdir.path())
-            .arg("extension")
-            .arg("install")
-            .arg(fixtures_path().join("module-import-extension").join("successful"))
-            .assert()
-            .success();
-
-        let cmd = Command::cargo_bin("phylum")
-            .unwrap()
-            .env("XDG_DATA_HOME", tempdir.path())
-            .arg("module-import-success")
-            .assert()
-            .success();
-
-        let stdout = std::str::from_utf8(&cmd.get_output().stdout).unwrap();
-        assert!(stdout.contains("I should contain 12345"));
-    }
-
-    // The fixture for this test attempts a directory traversal.
-    #[test]
-    fn module_with_traversal_fails_to_load() {
-        let tempdir = TempDir::new().unwrap();
-
-        Command::cargo_bin("phylum")
-            .unwrap()
-            .env("XDG_DATA_HOME", tempdir.path())
-            .arg("extension")
-            .arg("install")
-            .arg(fixtures_path().join("module-import-extension").join("fail-local"))
-            .assert()
-            .success();
-
-        let cmd = Command::cargo_bin("phylum")
-            .unwrap()
-            .env("XDG_DATA_HOME", tempdir.path())
-            .arg("module-import-fail-local")
-            .assert()
-            .failure();
-
-        let stderr = std::str::from_utf8(&cmd.get_output().stderr).unwrap();
-        assert!(stderr.contains("importing from paths outside"));
-    }
-
-    // The fixture for this test attempts to load a module from a non-`deno.land`
-    // URL.
-    #[test]
-    fn module_with_non_allowed_url_fails_to_load() {
-        let tempdir = TempDir::new().unwrap();
-
-        Command::cargo_bin("phylum")
-            .unwrap()
-            .env("XDG_DATA_HOME", tempdir.path())
-            .arg("extension")
-            .arg("install")
-            .arg(fixtures_path().join("module-import-extension").join("fail-remote"))
-            .assert()
-            .success();
-
-        let cmd = Command::cargo_bin("phylum")
-            .unwrap()
-            .env("XDG_DATA_HOME", tempdir.path())
-            .arg("module-import-fail-remote")
-            .assert()
-            .failure();
-
-        let stderr = std::str::from_utf8(&cmd.get_output().stderr).unwrap();
-        assert!(stderr.contains("importing from domains other than"));
-    }
-
-    // A symlink is directly created during the test, as no symlinks are committed
-    // to the repo.
-    #[cfg(unix)]
-    #[test]
-    fn symlinks_are_rejected() {
-        let tempdir = TempDir::new().unwrap();
-        let ext_path = tempdir.path().join("phylum").join("extensions").join("symlink-extension");
-
-        Command::cargo_bin("phylum")
-            .unwrap()
-            .env("XDG_DATA_HOME", tempdir.path())
-            .args(&["extension", "install"])
-            .arg(fixtures_path().join("symlink-extension"))
-            .assert()
-            .success();
-
-        std::os::unix::fs::symlink(ext_path.join("symlink_me.ts"), ext_path.join("symlink.ts"))
-            .unwrap();
-
-        Command::cargo_bin("phylum")
-            .unwrap()
-            .env("XDG_DATA_HOME", tempdir.path())
-            .arg("symlink-extension")
-            .assert()
-            .failure()
-            .stderr(predicate::str::contains("importing from symlinks is not allowed"));
-    }
 }
 
 ////////////////////////////////////////////////////////////////////////////////
